add multi-region to storage instance provisioning (#25)
add multi-region perf test infra scripts

+++ scripts/setup_env.sh
@@ -5,10 +5,17 @@ sudo apt-get -y upgrade
 
 sudo apt install make -y
 
-wget https://dl.google.com/go/go1.16.6.linux-amd64.tar.gz
-sudo tar -xvf go1.16.6.linux-amd64.tar.gz
-sudo mv go /usr/local
-rm -f go1.16.6.linux-amd64.tar.gz
+
+if [ -d "/usr/local/go" ] 
+then
+    echo "Go already installed" 
+    echo "skip installing go"
+else
+    wget https://dl.google.com/go/go1.16.6.linux-amd64.tar.gz
+    tar -xvf go1.16.6.linux-amd64.tar.gz
+    sudo mv go /usr/local
+    rm -f go1.16.6.linux-amd64.tar.gz
+fi
 
 echo 'export GOROOT=/usr/local/go' >> ~/.profile
 echo 'export GOPATH=$HOME/go' >> ~/.profile
